// src/app/interfaces/CargarUsuariosResponse.ts
// Generated by https://quicktype.io

export interface CargarUsuariosResponse {
  ok:             boolean;
  usuarios:       Usuario[];
  totalRegistros: number;
}

export interface Usuario {
  _id: string;
  nombre: string;
  email: string;
  role: string;
  google: boolean;
  img?: string;
  imgenUrl: string;
  imagenSetter:string;
}